Remove support for end-of-life Pythons
Match pip support Pythons and remove support for end-of-life Pythons.
pip has not support Python 2 since 21.0 (2021-01-23). For dates on when
these environments went EOL, see:

https://devguide.python.org/devcycle/#end-of-life-branches

Removing support for Python2 allows for several simplifications:

- Remove use of toml package in favor of tomli
- Remove compat.py
- Replace io.open with builtin open
- Replace tempdir function with tempfile.TemporaryDirectory
- Replace mkdir_p function with os.makedirs
- Remove use of mock package in favor of unittest.mock
- Remove import from future module
- Remove unnecessary inheritance from object

Refs #90
