+++ dev-requirements.txt
@@ -1,11 +1,8 @@
 pytest
 pytest-flake8
 flake8 < 4  # https://github.com/tholo/pytest-flake8/issues/81
-pytest-forward-compatibility; python_version<'3'
-mock ; python_version<'3.6'
 testpath
-toml ; python_version<'3.6'
-tomli ; python_version>='3.6'
+tomli
 setuptools>=30
 importlib_metadata ; python_version<'3.8'
 zipp ; python_version<'3.8'
