+++ doc/requirements.txt
@@ -1,2 +1,1 @@
-toml ; python_version<'3.6'
-tomli ; python_version>='3.6'
+tomli
