+++ pep517/colorlog.py
@@ -73,8 +73,6 @@ class LogFormatter(logging.Formatter):
             # right conversion in python 3.
             fg_color = (curses.tigetstr("setaf") or
                         curses.tigetstr("setf") or "")
-            if (3, 0) < sys.version_info < (3, 2, 3):
-                fg_color = str(fg_color, "ascii")
 
             for levelno, code in self.DEFAULT_COLORS.items():
                 self._colors[levelno] = str(
