+++ pep517/dirtools.py
@@ -1,31 +1,6 @@
-import contextlib
-import errno
 import io
 import os
-import shutil
-import tempfile
 import zipfile
-
-
-@contextlib.contextmanager
-def tempdir():
-    """Create a temporary directory in a context manager."""
-    td = tempfile.mkdtemp()
-    try:
-        yield td
-    finally:
-        shutil.rmtree(td)
-
-
-def mkdir_p(*args, **kwargs):
-    """Like `mkdir`, but does not raise an exception if the
-    directory already exists.
-    """
-    try:
-        return os.mkdir(*args, **kwargs)
-    except OSError as exc:
-        if exc.errno != errno.EEXIST:
-            raise
 
 
 def dir_to_zipfile(root):
