+++ pyproject.toml
@@ -9,15 +9,15 @@ author-email = "thomas@kluyver.me.uk"
 home-page = "https://github.com/pypa/pep517"
 description-file = "README.rst"
 requires = [
-    "toml;python_version<'3.6'",
-    "tomli >=1.1.0;python_version>='3.6'",
+    "tomli >=1.1.0",
     "importlib_metadata;python_version<'3.8'",
     "zipp;python_version<'3.8'",
 ]
+requires-python = ">=3.6"
 classifiers = [
     "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 2.7",
     "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3 :: Only",
 ]
 
 [tool.isort]
