+++ tests/test_hook_fallbacks.py
@@ -1,12 +1,11 @@
-import io
 from os.path import abspath, dirname
 from os.path import join as pjoin
 
 import pytest
+import tomli
 from testpath import assert_isfile, modified_env
 from testpath.tempdir import TemporaryDirectory
 
-from pep517.compat import toml_load
 from pep517.wrappers import HookMissing, Pep517HookCaller
 
 SAMPLES_DIR = pjoin(dirname(abspath(__file__)), 'samples')
@@ -15,8 +14,8 @@ BUILDSYS_PKGS = pjoin(SAMPLES_DIR, 'buildsys_pkgs')
 
 def get_hooks(pkg):
     source_dir = pjoin(SAMPLES_DIR, pkg)
-    with io.open(pjoin(source_dir, 'pyproject.toml'), 'rb') as f:
-        data = toml_load(f)
+    with open(pjoin(source_dir, 'pyproject.toml'), 'rb') as f:
+        data = tomli.load(f)
     return Pep517HookCaller(source_dir, data['build-system']['build-backend'])
