+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27, py34, py35, py36, py37, py38, py39, py310, pypy, pypy3, isort
+envlist = py36, py37, py38, py39, py310, pypy3, isort
 skipsdist = true
 
 [testenv]
